fix: remove intermediate step
BREAKING CHANGE: removes legacy import style {{import ...}}, which removes extra steps during build
