+++ .eslintrc.js
@@ -18,7 +18,6 @@ module.exports = {
   rules: {
     'ember/no-jquery': 'error',
   },
-  rules: {},
   overrides: [
     // node files
     {
