+++ package.json
@@ -64,7 +64,7 @@
     "ember-try": "^1.4.0",
     "eslint": "^6.8.0",
     "eslint-plugin-ember": "^7.10.1",
-    "eslint-plugin-node": "^11.0.0",
+    "eslint-plugin-node": "^11.1.0",
     "husky": "^1.3.1",
     "jest": "^24.1.0",
     "loader.js": "^4.7.0",
